requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
jinja2>=3.1.0

# AI Model Providers
anthropic>=0.40.0
openai>=1.55.0
google-genai>=1.0.0

# Math verification
sympy>=1.13.0

# Export
reportlab>=4.0.0
python-docx>=1.0.0

# Utilities
pydantic>=2.9.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
